#1709: split WG handshake orchestration into handshake_session.rs
The handshake integration (the four slow-path entry points, the
HandshakeError enum, the PendingHandshake reservation record, the
two-phase index reservation, and the rand index helper) pushed
engine.rs from 1725 to 2245 LOC, crossing the 2000-LOC [REFACTOR]
audit threshold the plan committed to staying under.

Move that orchestration into a new wg/handshake_session.rs. The methods
are `impl WgEngine` in the same `wg` module and reach the engine's
fields and `peer_arc` via `pub(in crate::afxdp::wg)` visibility (the
project's module/foo.rs split idiom). engine.rs keeps the WgEngine
struct, the peer table, the build_*_handshake snow builders, and the
hot encap/decap paths; it returns to 1772 LOC (WATCH-tier, under the
2000 threshold). No behavior change — pure code motion; the 102 wg
tests are unchanged and still pass.
